requirements.txt
django-simpleui
